+++ packages/site/src/content/InputEmail/index.tsx
@@ -1,0 +1,20 @@
+import { InputEmail } from "@jobber/components";
+import InputEmailContent from "@atlantis/docs/components/InputEmail/InputEmail.stories.mdx";
+import Props from "./InputEmail.props.json";
+import { ContentExport } from "../../types/content";
+
+export default {
+  content: () => <InputEmailContent />,
+  props: Props,
+  component: {
+    element: InputEmail,
+    defaultProps: { placeholder: "Enter your email" },
+  },
+  title: "InputEmail",
+  links: [
+    {
+      label: "Storybook",
+      url: "http://localhost:6006/?path=/docs/components-utilities-InputEmail-web--docs",
+    },
+  ],
+} as const satisfies ContentExport;